backend/requirements.txt
alembic==1.11.1
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
comm==0.1.4
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
executing==1.2.0
fastjsonschema==2.18.0
firebase-admin==6.2.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
fonttools==4.42.1
fqdn==1.5.1
gcloud==0.17.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.4
iniconfig==2.0.0
install==1.3.5
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
jws==0.1.3
kiwisolver==1.4.5
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.3
matplotlib-inline==0.1.6
mistune==3.0.1
msgpack==1.0.5
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.25.2
oauth2client==3.0.0
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipdeptree==2.13.0
platformdirs==3.10.0
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.0
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-json-logger==2.0.7
python-jwt==2.0.1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==0.7.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
Send2Trash==1.8.2
sendgrid==6.10.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.15
stack-data==0.6.2
starkbank-ecdsa==2.2.0
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.6.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
validate-email==1.3
validate-email-address==1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.4
